backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
openai
python-dotenv
pandas
numpy
torch

